LIT 6193 6576 6616 617 6618 6619
Some updates and a few new records around Dersana Madhane Alam. I created a record (6193) for versions that do not coincide with the one edited (more or less, it is 6193 - i change the title to dissimilate it from another collecrion). 6616 and 6617 are for pieces that circulate through two collections at least. 6617 is for a series of short homilies sometimes  appended to 6618 works