Create Bibliography to LIT4310Prayer (#960)
* Create Bibliography to LIT4310Prayer

Recorded bibliography according to the account of P. Marrassini in his catalogue of the Florentine manuscripts. I was not able to identify the following record: B. Turaev, VVr 15, 1909, pp. 183-5. I am not sure which title is concerned and cannot verify this record, that is not listed in Marrassinis catalogue.

* Update LIT4310Prayer.xml

Added edition of Turaev.