Create LIT6831PrayerGeorge
I created an ID for LIT6831PrayerGeorge as attested in BMLacq306. A very short prayer, that is very similar to the final part of LIT4518Hymn. Although very short altogether, I decided to create a new ID for this piece as discussed in issue #2319. I gave the entire text as a div.