Create LIT6832CommentaryApocalypseJohn
I created LIT6832CommentaryApocalypseJohn as it is attested in BMLacq38 and discussed in issue #2335.
I did not know, how to encode the relations to LIT3179Revela and PRS5695John, that this commentary is dealing with. I do not know, what can be written in an abstract except for what is already given in the title.