New LIT for Zəmmāre za-Hosāʿnā and Zəmmāre za-Dabra Zayt
I created two new ID for Zəmmāre za-Hosāʿnā and Zəmmāre za-Dabra Zayt according to our discussion in Issue #2380. I added a relation to point to the general record of LIT2610Zemmar as it was discussed to treat them as if they were part-texts of the general record.