Created LIT for Mawāśəʿt za-ʾArsimā (#1003)
* Created LIT for Mawāśəʿt za-ʾArsimā

I created a LIT ID for Mawāśəʿt za-ʾArsimā, as attested in BMLor148 and discussed in Issue #2381. I added this ID also to the general record as <div>, because quiet a good number of Antiphons is listed there as <div> already.

* Update LIT1990Mawase

I added xml:id to all corresp= as indicated by Eugenia in Issue #2381.