New LIT for Zəmmāre za-Hosāʿnā and Zəmmāre za-Dabra Zayt (#1015)
* New LIT for Zəmmāre za-Hosāʿnā and Zəmmāre za-Dabra Zayt

I created two new ID for Zəmmāre za-Hosāʿnā and Zəmmāre za-Dabra Zayt according to our discussion in Issue #2380. I added a relation to point to the general record of LIT2610Zemmar as it was discussed to treat them as if they were part-texts of the general record.

* Updated Zəmmāre records

I updated the titles and added a note to the newly encoded zəmmāre records of LIT6848ZemmareDabraZa and LIT6847ZemmareHosana.

* Update LIT6847ZemmareHosana

I changed the title according to our discussion in the previous pull request.