Update Tarika Walda Amid (#1059)
* Updated Tarika Walda Amid

I updated LIT4723TarikaWaldaAmid and created two more records for the two recensions (α&β). I added labels to the index of the textparts and decided to move them to LIT6936TarWaAmidBeta, because the description is based on Goldschmidts catalogue entry to a manuscript of the β-recension.

* Update records on Tarika Walda Amid

I updated the records according to Eugenias requests with the exception of links in other work records and in manuscript records, that I will care about in a minute.

* Update LIT records related to Tarika Walda Amid

I replaced the link to the Tarika Walda ʾAmid (LIT4723TarikaWaldaAmid) by a pointer to the Beta-recension (LIT6936TarWaAmidBeta). However, I am not perfectly content with that structure, because I think, that the different layers can be described better in another way.

* Update Tārika Walda ʾAmid records

I updated the records according to Dorotheas requests.

* Update LIT4723TarikaWaldaAmid.xml

---------

Co-authored-by: thea-m <thea.m@hotmail.de>