Update LIT1544Gebrah (add chapters as Sub-IDs) (#1101)
* Update LIT1544Gebrah

I added some chapters to the record of LIT1544Gebrah as discussed in https://github.com/BetaMasaheft/Documentation/issues/2531.
I also added some titles to chapters already existing.
However, I saw, that there is no subdivision for each hour of the day or of the night (except for Friday), as it is regularly in manuscripts. If you agree, I would like to add also chapters for the other days of the week, unless the place of some readings may vary in some manuscripts.

* Update LIT1544Gebrah

I updated div textparts according to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2531.