Delete records for Admonitions
I discovered that one of the admonitions (sena fetratem) is also listed in a catalogue of manuscripts in Cambridge (Ullendorff 1961, no. 43, f. 39a) as part of . It is therefore very likely that both admonitions are part of LIT4014Medan. I am sorry that I overlooked this in Strelcyns catalogue and created three unnecessary IDs.