Create LIT7041LordsPrayerAmh
I created a record for the Lord's Prayer in Amharic as attested in BLorient7946. I was wondering, whether I should name "Lord's prayer (Amharic)" more prominently in the title, but I decided to use the Amharic title to be consistent with our rule of how to encode prayers in Bm.