Create three qene
I am not content with the titles of LIT7060Qene (also no accusative in mǝśyāṭ) and LIT7062Qene, that do not make sense to me. I hope, you will find better solutions.

I am not experienced enough in Qene to identify the qene type with certainty. LIT7060Qene may be of Zayǝʾǝze-type as well, because it has 5 verses.