Update LIT1763Legend.xml (#1184)
* Update LIT1763Legend.xml

I updated the bibliography with additional literature, that I found in Strelcyn 1978 for BLorient11602.

However, I have seen, that some of the existing literature in the bibliography was not well displayed, presumably because the tag bm:Worrell1886Zauberwesen does not exist in Zotero. I did not find any publication of 1886 with the name Worrell in KVK, as well. For this reason, I think, it was a mistake that I corrected to bm:Worrell1909Zauberwesen and bm:Worrell1910Zauberwesen

Worrells publication of 1909 has both edited texts and translations, while Worrell 1910 contains only translated texts from various mss.

* Update LIT1763Legend.xml