Correct LIT4305Prayer
There are two LIT ID in BAVet36 with the same title:

LIT4304Prayer -> Prayer for obtaining freedom from the enemies
LIT4305Prayer -> Prayer for obtaining freedom from the enemies

However, the second one is described as 'Prayer for having children' in the manuscript record of BAVet36. I do not understand from the given text in the incipit and the explicit, what it is about, but I see in the catalogue (GrébautTisserant1935, p. 167), that the latter is correct: 'Precatio ad prolem obtinendam' = "A prayer to obtain a child".

I think the title of the record LIT4305Prayer is wrong because of copy&paste from the previous one. I corrected it. For comparison see the manuscript record of BAVet36#ms_i12:

 ```
 <msItem xml:id="ms_i12">
                     <locus from="115r" to="116r" facs="0130"/>
                     <title type="incomplete" ref="LIT4305Prayer">Prayer for having children</title>
                     <incipit xml:lang="gez">አምላክየ፡ <sic>ለዓለም፡</sic>ዓለም፡ <sic>ነገሢ፡</sic> አሜን፡ ወተዓትብ፡ <sic>ገጽከ፡</sic> ሰብዓቱ፡ ጊዜ፡ <sic>ጸውዕ፡</sic><sic>ኤልሰዕ፡</sic> ለሰምራዊት፡ ወይቤል፡ ወቆመ፡ ኀበ፡ ሖሕት፡  </incipit>
                     <note>The beginning is missing.</note>
                     <explicit xml:lang="gez"> ወእምድሕረ፡ ሰ<gap reason="lost" extent="unknown" resp="PRS4805Grebaut PRS9530Tisseran"/> ወሰከቡ፡ <sic>ኢይትነገሩ፡</sic><gap reason="lost" extent="unknown" resp="PRS4805Grebaut PRS9530Tisseran"/> ወልድ፡ በፈቃደ፡ እግዚ<supplied reason="lost">አብሔ</supplied>ር፨ ሎቱ፡ ይደሉ፡ ስ<supplied reason="lost">ብሐ</supplied>ት፡ ለዓለመ፡ ዓለም፡ <supplied reason="lost">አ</supplied>ሜን፨</explicit>
                  </msItem>
```