Update LIT4145Onausp (#1149)
* Update LIT4145Onausp

I updated the titles and added literature that I found in Strelcyns catalogue on BLorient9897.

* Update LIT4145Onausp (languageDesc)

* Update LIT4145Onausp

According to Jonas remarks in https://github.com/BetaMasaheft/Manuscripts/pull/2511.

* Update LIT4145Onausp according to Denis requests

I checked four manuscripts, that are given by Gasparini. They all have similar incipits and therefore are likely to have the same text. I added them in the listWit, except for BnF d'Abbadie, that does not exist as a stub, yet. Do you want me to create a stub for it and add it to the list?

* Update LIT4145Onausp according to Denis requests

* Update LIT4145Onausp according to Eugenias request