Create LIT7094PrayerDevils (#1234)
* Create LIT7094PrayerDevils

I created this record following our discussion in https://github.com/BetaMasaheft/Documentation/issues/2622. Contrary to what I envisaged in the issue, I decided to mark the similar Asmāt in other works or manuscripts.
This Prayer also exhibits an unusual feature as it has a supplication በስመ፡ በስሙ፡ ለአብ፡ በስሙ፡ ለወልድ፡ በስሙ፡ ለመንፈስ፡ ቅዱስ፡ which is slightly different from what is found in most prayers (በስመ፡ አብ፡ ወወልድ፡ ወመንፈስ፡ ቅዱስ፡). However, I do not know, whether there is something specific or noteworthy on this particular supplication in the incipit.

* Update LIT7094PrayerDevils

Added bibliography and updated translation.