Create LIT7103PrayOTprophets
I created a provisional record for LIT7103PrayOTprophet according to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2647.

According to my investigation, the BnF manuscripts in the listWit do not have any records so far. If I am not mistaken and if they are not processed in any pull requests at the moment, I want to create stubs for them.