Amharic admonitions (meedan) in BLorient11719 ch july24 (#1276)
* Created LIT records for Admonitions in BLorient11719

Following my suggestions in
https://github.com/BetaMasaheft/Documentation/issues/2566.

* Delete records for Admonitions

I discovered that one of the admonitions (sena fetratem) is also listed in a catalogue of manuscripts in Cambridge (Ullendorff 1961, no. 43, f. 39a) as part of . It is therefore very likely that both admonitions are part of LIT4014Medan. I am sorry that I overlooked this in Strelcyns catalogue and created three unnecessary IDs.

* Revert "Delete records for Admonitions"

This reverts commit 1ac1c2f7a503a23bc3a4bb585f189305b67e47dc.

* Revert the delete commit.

I reverted the deletion as you said. However, I was not able to push any pull request, because the github desktop answered, that I have made no changes. I did some small changes in order to be able to commit to this branch.