Create LIT7125QeneRufael (Qene from BNFabb145, no. 10)
I created a record for the Qene of folio 55vb-c. I was challenged by the translation of the first line, whose content I do not understand. I am wondering, whether twice ተሰብኦ might be a dittography.

Furthermore, I have problems in determining the type. The poem is divided with verse markers (።) into six or seven verses depending on the emendation in line 4. However, the rhyme in ለ indicates, that the first verse break must be a mistake, because otherwise the first verse would end in ጸበተ and would not match to the rhyme in ለ (accepting the crossing out of ፩ at the end of the first line. Otherwise, there could be also a rhyme in ጸበተ and  ፩ (= ʾaḥatta), what is difficult to accept, because a change in the rhyme from ተ to ለ is not common in qene, as far as I know.) Another option could be to define two very long verses ending in ብህለ (if this emendation is correct) and መሰለ and a kwellekomu or gubāʾe qānā type respectively.