Update records concerning prayers against stomachache
I have seen that LIT4624Prayer and LIT6716PPWegat are essentially the same prayer with the only difference, that LIT6716PPWegat contains ʾasmat. I wanted to indicate this relationship by adding relation in body and by updating the abstracts.
I do not know, what happened to LIT4705Prayer, which was mentioned in line 65 of LIT4624Prayer. LIT4705Prayer is not in the app. I think, this note can be deleted.