Create LIT14518QeneOzyan.xml
I created a record for another Qene in BNFabb145. I was not certain about the number of verses, because one of them in line 5 is not visible even in the pictures, which Nafisa has provided. On the other hand, there are two words in line 1 and 2, which end in -lā, as it is perfectly the case with all the rhymes in the poem. I am not sure, if these could be regarded as verse endings as well and the poem having 9 or even 10 verses in the end. With 9 verses, it would be a kwellekemu type, but a type with 10 verses is not mentioned in Guidi, if I read it correctly.
The name ʿOzyān (=Uzziah) is prominent also in LIT7082QeneEmnahsa, LIT1650Homily, LIT2671CanonsSimon2,
LIT2639CanonsSimon1 and LIT6936TarWaAmidBeta. I would like to create a PRS ID for him and update the LIT records mentioned. I think, there is no PRS record for Uzziah yet and he is prominent enough in Ethiopian literature to get one.