Update LIT1295Dersan
According to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2673.
I decided to take the account of BMLacq776 as basis for my edition, because I have extensive incipits and texts for this one and it is available online. Beyond that, I do not have any reason to prioritize any of the manuscripts mentioned in the record. However, the order and attribution of the homilies and miracles can vary considerably in other witnesses. So, I decided to delete many references to months in labels and renamed many of them.