Create LIT7239Miscarriage.xml
I created an ID for the prayer, which I described and discussed in https://github.com/BetaMasaheft/Documentation/issues/2768.

Curiously, I do not find any hint to 'vaginal haemorrhage' and 'miscarriage' that Strelcyn claims this prayer to be about. These 'women's diseases' may be expressed more politely with ዓምደ፡ ብርሃን፡ መስረተ፡ ጽድቅ፡, which I translated with "Pillar of the light, foundation of rightousness".