Created LIT7240PPAddam
I created a record for LIT7240PPAddam as discussed in https://github.com/BetaMasaheft/Documentation/issues/2769.
I had difficulties to explain the term ዘለለሐከ. In cannot find ለሐከ in the dictionaries. I thought, it may mean "revealed". If we cannot verify this meaning, I would better not take it for the title.

I updated the title in LIT7233PrColic.