Update labels textparts according edition Diez
I decided to update the labels of the textparts according to the edition of the Arabic text as published by Diez, which covers the text from Adam to the end of the Achaemenids. I see the transliterations of the Ethiopic forms, which stem from ms. J (FSUor134) not useful, because many cases are random garbles and the orthography is not representative for most witnesses. Also the enumeration is distorted and should be based on the Arabic Vorlage as soon as the second part of the edition is published by Diez/Muccini later this year. For this reason, I deleted the numbers of the generations for the latter parts.