Create LIT7237QeneFequrana.xml (#1389)
* Create LIT7237QeneFequrana.xml

Several features are problematic in this Qene in BNFet145. First problem is the verse break after መፍርህ, disturbing the rhyme, which is always in ል in this poem. For this reason, I deleted this verse break and removed it to ሣህል, what is two words before.
The second problem is the addition of ጊዜ፡ ሞቶሙ፡ ረድኤት፡  and ወ, which was inscribed between the lines. I am not completely sure, whether I inserted these words on the correct place in verse 1. You may check, if this verse sounds smoothly to you.
The third is issue is the translation of the title. I translated it as circumstantial clause similar to Arabic ḫāl-sentence. You may have another proposal. Possibly, it may be better to include more words from the first verse to the title.
Fourth is the type. If I understand Guidi correct, every eight verses are mawaddes and nine verses are mawaddes kwellekemu. It sounds not very logical to me and my Italian is too weak to be sure about Guidis statement. Please check and approve my assumption (also relevant for LIT7223QeneMenase and other records of mawaddes kwellekemu).
Fifth I did not follow Nafisas reading ኢሎፍሲ፡ and preferred ኢሎፍሊ፡, because I think it is about 'Philistines'.

* Update according to Denis and Jonas requests