Create LIT7275QeneWaEsagged.xml
I created a new LIT for a qene in BNFet145 (f. 56rb).

I was confused about the ዋ። in the beginning. I interpreted it as and interjection similar to "Oh" in English. I have never seen such an interjection or particle in Geez. I did not counted it in the counting of verses but I interpreted it as a part of the text.