Create LIT7279QeneAlbeki.xml
I do not understand the meaning of the first verse and hope, that my translation of the title will make some sense. I am aware, that biblical Shiloh was located in Judea, but not in Lebanon. Perhaps a confusion with Salem/Jerusalem which is mentioned in verse 8 - however with the same problem. I will be grateful for other suggestions!

With the intervention of adding one more verse break in line 5 I come up with 9 verses resulting in kwellekemu-type. Otherwise there are eight lines in this qene resulting in a mawaddǝs-type. What do you think - is it legitimate to make this emendation? If not divided in two, the verse no. 5 would be very long.