Create LIT7301QeneBalelit.xml
I created a record for a qene found in BNFabb145 (the last one in this collection).
In my edition, this is a qene of the mawaddǝs-kʷǝllǝkǝmu-type having nine verses. However, the insertion of three more verse breaks after words ending in ሎ was necessary to achieve this number. Without these emendations, the number of verses would have been six ending up in a śǝllāse-type.
I do not know the story, that this qene is referring to. I identified Moʿāb with the biblical Moab and translated Barbara with 'Barbaria', but I am not quite sure about the latter.
I also emended ያጠፍእ to ያጠፍኡ. Would you please check the translation and the text following after and give me a note, whether you think these emendations are appropriate and correct?