Create LIT7324QeneIfatara
I created a record for a qene from EMIP00764, which Denis recommended to me and which is mentioned in ...

The type is indicated above with Śāhlǝka as I understand Denis request in a facebook message. However, this type is not mentioned in Guidi 1901 and we do not have it as an authFile and should possibly create one.

The text is written in a crude hand as an addition on the upper margin. One character is partly lost and I needed to restore it to ሙ to achieve three verses rhymed in ሙ. I hope, you will agree with this emendation.

The content seems to be very abstract and panegyric. You may have a look on the translation and on the edition of the text and may help me with some improvement.