Create LIT14520QeneZawahabkani.xml (#1504)
* Create LIT14520QeneZawahabkani.xml

I created record for a Qene poem, which Denis has directed me to. However, the creation of the ID through the app ended up with the unhappy number of LIT14520, what is beyond the expected number in the 7300s. I suggest to change the number, before the branch may be eventually merged.

I did not find an authority file for the Qene-type Za-ʾamlākiya and pronounced the wish to create one in https://github.com/BetaMasaheft/Documentation/issues/2504 and  https://github.com/BetaMasaheft/Documentation/issues/2830.

For the end of the poem, there is one more phrase, which I think, does not belong to the poem (ተፈጸመ፡ ግብረ፡ ሰዓታት።), but may be some instruction or indication of the end of the service or something similar. I did not include this line to the edition of the text.

* Update according to Jonas remarks

* Create LIT7319QeneZawahab

* Up folio number 157rb

I have discovered, that the newly recorded qene is attested on folio 157rb (not: 160r).

* Updates and add translation according to Denis requests

* Update bestow according to Denis remark