Create LIT7336QeneMaryam, update LIT7048Qene
According to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2856.

I had difficulties to translate the first line in each of the poems and kept the titles short. If you can provide a nice translation for the first verse, I would be happy to update the title in each case.