Create LIT7346QeneLasabaya.xml
I created a new LIT ID for a qene in PetermannIINachtrag24 as suggested in https://github.com/BetaMasaheft/Documentation/issues/2830#issuecomment-2802066886.

This Qene has 9 verses and is indicated as Mawaddes. It should be thus classified as kʷǝllǝkǝmu mawaddǝs according to Guidis classification.

However, I found the same situation in LIT7279QeneAlbeki, which was classified just as kʷǝllǝkǝmu and which must probably be updated to kʷǝllǝkǝmu mawaddǝs, as well.

My translation of the title is rather hazard and may possibly need some improvements.