Create LIT7496Sawwer.xml
According to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2998.

The title and the incipit was not very informative to me. It seems, that there is a lot of astrological or magical terminology, that I cannot understand or translate. For this reason, I preferred to keep Kalkalos, Zāru and Ṭimos as proper names. Between these enigmatic terms, there are some words (especially የሚባል), that remind to Amharic, and some, which are Gəʾəz. I identified it therefore as a Gəʾəz-text, which is "partly Amharic".
I do not know, what the term sawwər (= conceal/hide, Imperative) stands for. It might be the case, that we are informed about conjunctions, whereas one planet will 'conceal' another or any other magic meaning of the word.