Title modified (#1591)
* Title modified

The manuscript tradition of Tansǝʾa Wald and Maʿāzā Dǝngǝl, respectively the founding and deputy abbots, reflects their close spiritual and devotional connection. Although the Clavis ID is technically applicable as it stands, I have made a slight modification to the title of the work, since it is traditionally known as a combined hagiography. The roles, miracles, and sälǝmāt of the two monks are, in all known cases, compiled together in a single manuscript that narrates the lives of both saints. While the text predominantly offers a detailed biography of ʾAbbā Tansǝʾa Wald, the founder, it presents their stories jointly. Moreover, the church preserves the tabots of both monks, further affirming their shared sanctity and veneration.

* Update LIT1076GadlaTensaeWald.xml