Update Excerpts of Walda Amid
I am not content with two records related to Tārika Walda ʾAmid. These texts, which are edited by Kropp 2016 do not contain the introductory chapters of the Tārika Walda ʾAmid, but contain excerpts from nearly all parts of the chronicle. They form the introduction to later chronological works. The titles 'Introduction of Walda ʾAmid' and 'Abbreviated Introduction of Walda ʾAmid' is therefore confusing and should be changed. I see them as excerpts from the Tārika Walda ʾAmid and added therefore saws:isShorterVersionOf as relations.
As well, the reference to the Beta-recension (LIT6936TarWaAmidBeta) is not justified. As far as I can see, the relation to any specific recension is not established yet. Therefore, we should stay with the general LIT4723TarikaWaldaAmid.