Update LIT1927Mashaf
I added abstract, keyword and edition to LIT1927Mashaf (Maṣḥafa Faws). To my surprise, I could not find an article about the Maṣḥafa Faws (Treatise of Therapeutics) in EAe, vol. 3 and 4. Possibly, it is among the Addenda in vol. 5, which is currently not available to me. I will check it, when I am in Hamburg again.