Update LIT1830Mahbar.xml
I updated LIT1830Mahbar with information from EAe and from the catalogue description by St. Strelcyn for BLorient13309 including a lengty incipit. In this manuscript, the title t1 is explicitly mentioned, while t2 is absent. In the EAe article by Samuel Yalew, both are given without priority. I decided therefore to prioritise t1 () over t2 as I assume, that this is the original title and t2 may be a secondary title.