Update LIT1931Mashaf.xml
I proposed some enhancements to LIT̀931Mashaf in https://github.com/BetaMasaheft/Manuscripts/pull/3142#event-20452299527, which I hope you agree.

I created some sub IDs for textparts, which frequently form part of Maṣḥafa Gǝnzat and which I could find in ANLcr71 and in at least one more manuscript. Some of them have already been referenced in the record in relations.
Second, I added many incipits from ANLcr71. Even though, I do not know, whether ANLcr71 should have any superiority over other witnesses or not and is a more random choice, I think, it is an asset to enrich the record with more Ethiopic texts, which make it easier to discover and unveil textparts in other manuscripts and eventuell develop the record further and possibly disambiguate versions and recension of Maṣḥafa Gənzat in the future. Possibly it was not necessary to mention the presence or non-presence of each single textpart in note. If this is an overload, I can delete or move to comments, as you wish.