Updated miracles of ZB
I updated the miracles of Zarʾa Buruk according to the edition by Jäger (and the translation by Ricci). Additional miracles as known from ANLcr83 and ANLcr115 are added after and their origin is indicated in a note.

I will update ANLcr83 and ANLcr115 in a minute to accomodate the new (= old) order of miracles.