Update LIT6353MartyrdomIschyrion.xml
I updated the list of witnesses to be conform with ANLcr134, which I prepare for update and I want to provide the names of the manuscripts by mentioning the place, library and collection as I think, it is correct. I got in problems with EMML6834, for which the repository is not entirely clear.